pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "google-generativeai>=0.8.5",
    "gunicorn>=23.0.0",
    "llama-index>=0.12.35",
    "llama-index-embeddings-gemini>=0.3.2",
    "llama-index-llms-gemini>=0.4.14",
    "psycopg2-binary>=2.9.10",
    "pymupdf>=1.25.5",
    "python-dotenv>=1.0.0",
    "werkzeug>=3.1.3",
]
